add show longitute & latetite in the api

+++ devices/serializers.py
@@ -1,8 +1,7 @@
-from django.contrib.auth.models import User, Group
+
 from django.core.exceptions import ObjectDoesNotExist
-from django.utils.datetime_safe import datetime
 from rest_framework import serializers
-from rest_framework.fields import UUIDField, IntegerField, DateTimeField, CharField
+from rest_framework.fields import UUIDField, IntegerField, DateTimeField, CharField, FloatField
 
 from devices.models import Device, Brightness
 
@@ -18,12 +17,16 @@ class DeviceSerializer(serializers.Serializer):
     uuid = UUIDField()
     datetime = DateTimeField()
     value = IntegerField(max_value=None, min_value=None)
+    latitude = FloatField(read_only=True)
+    longitude = FloatField(read_only=True)
     status = CharField(max_length=None, min_length=None, allow_blank=True, read_only=True)
 
     extra_kwargs = {
         'uuid': {'write_only': True, 'required': True},
         'datetime': {'write_only': True, 'required': True},
         'value': {'write_only': True, 'required': True},
+        'latitude': {'read_only': True},
+        'longitude': {'read_only': True},
         'status': {'read_only': True}
     }
 
@@ -37,32 +40,40 @@ class DeviceSerializer(serializers.Serializer):
             device = None
 
         if device:
-            brightness = Brightness.objects.create(
-                device=device,
-                datetime=validated_data['datetime'],
-                value=validated_data['value']
-            )
-            brightness.save()
 
-            brightness_list = Brightness.objects.filter(device=brightness.device,
-                                                        datetime__year=brightness.datetime.year,
-                                                        datetime__month=brightness.datetime.month,
-                                                        datetime__day=brightness.datetime.day).exclude(
-                id__in=[brightness.id, ])
+            if validated_data['value'] >= 0:
+                brightness = Brightness.objects.create(
+                    device=device,
+                    datetime=validated_data['datetime'],
+                    value=validated_data['value']
+                )
+                brightness.save()
 
-            for brightness_item in brightness_list:
-                # counter
-                brightness.counter = brightness.counter + brightness_item.counter
-                # value
-                brightness.value = brightness.value + (brightness_item.counter * brightness_item.value)
-                brightness.value = brightness.value / brightness.counter
+                brightness_list = Brightness.objects.filter(device=brightness.device,
+                                                            datetime__year=brightness.datetime.year,
+                                                            datetime__month=brightness.datetime.month,
+                                                            datetime__day=brightness.datetime.day).exclude(
+                    id__in=[brightness.id, ])
 
-                # delete old entry
-                brightness_item.delete()
+                for brightness_item in brightness_list:
+                    # counter
+                    brightness.counter = brightness.counter + brightness_item.counter
+                    # value
+                    brightness.value = brightness.value + (brightness_item.counter * brightness_item.value)
+                    brightness.value = brightness.value / brightness.counter
 
-            brightness.save()
+                    # delete old entry
+                    brightness_item.delete()
 
-            validated_data['status'] = "Success: data saved"
+                brightness.save()
+
+                validated_data['status'] = "Success: data saved"
+            else:
+                validated_data['status'] = "Show device information"
+
+            # add longitute & latitute
+            validated_data['latitude'] = device.location_lat
+            validated_data['longitude'] = device.location_lon
 
             return validated_data
         else:
